OnGoing : organizacao da estrutura 1
