--- deprecated/requirements.txt
@@ -1,3 +1,0 @@
-boto3==1.26.141
-invoke==2.2.0
-cookiecutter==2.5.0
