+++ pytrobot/core/feature/subprocess.py
@@ -77,3 +77,19 @@ class SubprocessManager(metaclass=Singleton):
                 print(f"Process {process_id} is no longer active.")
         else:
             print(f"No process found with ID {process_id}.")
+
+
+# Decorator
+def Subprocess(comando, captura_saida=True, captura_erro=True):
+    """
+    Função para acessar o SubprocessManager do PyTRobot e executar um subprocesso.
+    
+    Args:
+        comando (list): Lista de strings contendo o comando a ser executado.
+        captura_saida (bool, opcional): Indica se a saída padrão do subprocesso deve ser capturada (padrão: True).
+        captura_erro (bool, opcional): Indica se a saída de erro do subprocesso deve ser capturada (padrão: True).
+
+    Returns:
+        dict: Resultado do subprocesso, conforme retornado pelo SubprocessManager.
+    """
+    return PyTRobot.set_subprocess(comando, captura_saida, captura_erro)
