+++ pytrobot/scaffold/{{cookiecutter.project_name}}/tests/test_perfomer_state.py
@@ -1,5 +1,5 @@
 import unittest
-from pytrobot.core.dataset_layer import TransactionData
+from pytrobot.core.strategies.state.dataset_layer import TransactionData
 from {{cookiecutter.project_name}}.{{cookiecutter.project_name}}.src.performer_state import PerfomerState  # type:ignore
 
 # Classe de teste unitário
